pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "deckadence"
version = "0.1.0"
description = "Deckadence: highly visual slide deck designer, player, and exporter."
authors = [{ name = "Deckadence", email = "devnull@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm",
    "pydantic>=2",
    "pydantic-settings>=2",
    "httpx>=0.27",
    "Pillow>=10",
    "ffmpeg-python>=0.2",
    "nicegui>=1.4.0",
    "aiofiles>=23",
    "tenacity>=8",
    "click>=8",
    "ruamel.yaml>=0.18",
]

[project.scripts]
deckadence = "deckadence.cli:main"
